Mapping add 19/10/2022. No change to tests
