+++ common/src/main/java/com/dmdev/entity/presets/Role.java
@@ -1,0 +1,5 @@
+package com.dmdev.entity.presets;
+
+public enum Role {
+    USER, ADMIN
+}
