--- general/src/main/java/dmdev/entity/Person.java
@@ -1,28 +1,0 @@
-package dmdev.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Person {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String name;
-
-    @Column(name = "patr_name")
-    private String patrName;
-    private String surname;
-    private String picture;
-    private Integer importance;
-    private Integer origin;
-}
