--- general/src/main/java/dmdev/entity/Role.java
@@ -1,5 +1,0 @@
-package dmdev.entity;
-
-public enum Role {
-    USER, ADMIN
-}
