--- general/src/main/java/dmdev/entity/Tag.java
@@ -1,25 +1,0 @@
-package dmdev.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Tag {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    private String name;
-    private String description;
-    private Integer importance;
-    @Column(name = "color_hash")
-    private String colorHash;
-}
