--- general/src/main/java/dmdev/entity/profiles/PrivateProfile.java
@@ -1,22 +1,0 @@
-package dmdev.entity.profiles;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import java.time.LocalDate;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class PrivateProfile {
-
-    @Id
-    private Integer id;
-    private LocalDate wedding;
-}
