--- general/src/main/java/dmdev/entity/profiles/ProfessionalProfile.java
@@ -1,26 +1,0 @@
-package dmdev.entity.profiles;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.time.LocalDate;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class ProfessionalProfile {
-
-    @Id
-    private Integer id;
-    private String position;
-    private LocalDate hired;
-    private Integer scope;
-}
