--- general/src/test/resources/hibernate.cfg.xml
@@ -1,17 +1,0 @@
-<?xml version='1.0' encoding='utf-8'?>
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-<hibernate-configuration>
-    <session-factory>
-        <property name="connection.url">jdbc:postgresql://localhost:5432/test-storyteller</property>
-        <property name="connection.username">postgres</property>
-        <property name="connection.password">Masterhost25</property>
-        <property name="connection.driver_class">org.postgresql.Driver</property>
-        <property name="hibernate.dialect">org.hibernate.dialect.PostgreSQL10Dialect</property>
-        <property name="show_sql">true</property>
-        <property name="format_sql">true</property>
-<!--        <property name="hbm2ddl.import_files">dbTestInit.sql</property>-->
-<!--        <property name="hibernate.hbm2ddl.auto">create-drop</property>-->
-    </session-factory>
-</hibernate-configuration>
